よき

+++ src/getcc.js
@@ -2,6 +2,7 @@ const moment = require('moment');
 const training = require('./training.js');
 const generatecc = require('./generatecc.js')
 const sqlite3 = require("sqlite3");
+const { text } = require('express');
 const db = new sqlite3.Database("./main.db");
 
 
@@ -14,19 +15,20 @@ exports.getcolorcodec = function (message) {
     const colorcode = message.content.substr(1, 6);
 
     db.get(`SELECT * FROM data WHERE guildId = ${guildId} AND channelId = ${channelId}`, (err, row) => {
-        if (row.gamemode == 'training') {
-            training.trainingc(message);
-            
-        } else if (row.gamemode == 'oneshot') {
-            db.each(`SELECT * FROM oneshot WHERE guildId = ${guildId} AND channelId = ${channelId} AND userId = ${userId}`, (err, row2) => {
-                db.run(`DELETE FROM oneshot WHERE id = ${row2.id}`);
-            });
-            db.run(`INSERT INTO oneshot(date, guildId, channelId, userId, userName, colorcode) \
+        if (row) {
+            if (row.gamemode == 'training') {
+                training.trainingc(message);
+
+            } else if (row.gamemode == 'oneshot') {
+                db.each(`SELECT * FROM oneshot WHERE guildId = ${guildId} AND channelId = ${channelId} AND userId = ${userId}`, (err, row2) => {
+                    db.run(`DELETE FROM oneshot WHERE id = ${row2.id}`);
+                });
+                db.run(`INSERT INTO oneshot(date, guildId, channelId, userId, userName, colorcode) \
                 VALUES("${date}", ${guildId}, "${channelId}", "${userId}", "${userName}", "${colorcode}")`);
-            message.react('🤔')
+                message.react('🤔');
+            }
         } else {
-            generatecc.generateImage(colorcode)
-            message.channel.send({ files: [`./images/${colorcode}.png`] });
+            generatecc.generateImage(message, colorcode)
         }
     });
 }
